make a main folder for main components
