+++ client/src/components/Common/UserBoard.js
@@ -3,7 +3,7 @@ import React, { Component } from 'react';
 class UserBoard extends Component {
   render() {
     return (
-      <div className='user-board'>
+      <div className="user-board">
         <i className="fas fa-poll"></i>
         <span>Projects</span>
       </div>
