+++ client/src/index.js
@@ -1,15 +1,16 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './index.css';
+
 import App from './App';
 import * as serviceWorker from './serviceWorker';
-import { Provider } from "react-redux";
+import { Provider } from 'react-redux';
 import store from './store/store';
 
 ReactDOM.render(
   <Provider store={store}>
     <App />
-  </Provider>, document.getElementById('root')
+  </Provider>,
+  document.getElementById('root')
 );
 
 // If you want your app to work offline and load faster, you can change
