+++ package.json
@@ -13,7 +13,6 @@
   "axios": "^0.20.0",
   "bcrypt": "^5.0.0",
   "big-time": "^2.0.1",
-  "bulma": "^0.9.1",
   "connect-mongo": "^3.2.0",
   "cookie-parser": "^1.4.5",
   "cors": "^2.8.5",
